third_party/blink/web_tests/animations/keyframes-missing-arguments.html

<!DOCTYPE html>
<title>CSSKeyframesRule functions missing arguments</title>
<style>
  @keyframes anim {
  }
</style>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script>
  setup({single_test: true});
  var keyframes = document.styleSheets[0].rules[0];
  assert_class_string(keyframes, "CSSKeyframesRule");
  assert_throws_js(TypeError, () => { keyframes.appendRule(); });
  assert_throws_js(TypeError, () => { keyframes.deleteRule(); });
  assert_throws_js(TypeError, () => { keyframes.findRule(); });
  done();
</script>
